@@ -1,8 +1,8 @@
 package com.merca.back.dto;
 
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import javax.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.Setter;
