@@ -1,20 +1,21 @@
 package com.merca.back.model;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
-import javax.persistence.CascadeType;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.JoinTable;
-import javax.persistence.ManyToMany;
-import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.JoinTable;
+import jakarta.persistence.ManyToMany;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
 import javax.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
@@ -51,7 +52,7 @@ public class Ropa implements Serializable {
         joinColumns = @JoinColumn(name = "ropa_id"),
         inverseJoinColumns = @JoinColumn(name = "color_id"))
     private List<Color> colores = new ArrayList<>();
-    @Fetch(FetchMode.JOIN)
+    @Fetch(FetchMode.SUBSELECT)
 //    @ElementCollection(fetch = FetchType.LAZY)
 //    @JsonManagedReference
     @OneToMany(mappedBy = "ropa", cascade = CascadeType.ALL, orphanRemoval = true)
