@@ -1,12 +1,12 @@
 package com.merca.back.security.entity;
 
 import com.merca.back.security.enums.RolNombre;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import javax.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
