@@ -1,11 +1,9 @@
 package com.merca.back.service;
 
-import com.merca.back.model.Color;
 import com.merca.back.model.RopaColor;
 import com.merca.back.repository.RopaColorRepository;
 import java.util.List;
-import java.util.Optional;
-import javax.transaction.Transactional;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
